Remove old code and refactor in new code
