+++ app/src/main/AndroidManifest.xml
@@ -23,9 +23,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Poste"
         tools:targetApi="31">
-        <activity
-            android:name=".activities.Login2Activity"
-            android:exported="false" />
+
         <activity
             android:name=".activities.EditFolderActivity"
             android:exported="false" />
